:memo: New postdoc position at Byrd Polar
More than a month overdue, but time to move on with my new life post-PhD in a different hemisphere! Re-evaluating what I want to do, is it really Glaciology, or is it just Geography? Definitely sticking with the Deep Learning path. Excited to see where I will go to next, be it in academia or industry ;)