:rocket: Switch focus from data-driven to composable accelerated hardware
Data-driven research is so last decade, so re-focusing on composable data systems and accelerated hardware! Less about doing the research myself, but boosting others (linked to a good paper on that).

Still, need to care about reproducibility (ala conda-forge). Sunsetting zen3geo (for now) in favour of cog3pio, as Rust seems the way to go now (need to focus on low-level stuff for a while before going back high-level). Also removed link to website, people will know where to find me if they want to.